+++ packages/client-sdk-nodejs/src/internal/leaderboard-data-client.ts
@@ -40,6 +40,7 @@ import {
   Middleware,
   MiddlewareRequestHandlerContext,
 } from '../config/middleware/middleware';
+import {grpcChannelOptionsFromGrpcConfig} from './grpc/grpc-channel-options';
 
 export const CONNECTION_ID_KEY = Symbol('connectionID');
 
@@ -81,6 +82,8 @@ export class LeaderboardDataClient implements ILeaderboardDataClient {
     // index variable.
     this.nextDataClientIndex = 0;
 
+    const channelOptions = grpcChannelOptionsFromGrpcConfig(grpcConfig);
+
     this.clientWrappers = range(numDataClients).map(
       () =>
         new IdleGrpcClientWrapper({
@@ -88,23 +91,7 @@ export class LeaderboardDataClient implements ILeaderboardDataClient {
             new leaderboard.LeaderboardClient(
               this.credentialProvider.getCacheEndpoint(),
               ChannelCredentials.createSsl(),
-              {
-                // default value for max session memory is 10mb.  Under high load, it is easy to exceed this,
-                // after which point all requests will fail with a client-side RESOURCE_EXHAUSTED exception.
-                'grpc-node.max_session_memory':
-                  grpcConfig.getMaxSessionMemoryMb(),
-                // This flag controls whether channels use a shared global pool of subchannels, or whether
-                // each channel gets its own subchannel pool.  The default value is 0, meaning a single global
-                // pool.  Setting it to 1 provides significant performance improvements when we instantiate more
-                // than one grpc client.
-                'grpc.use_local_subchannel_pool': 1,
-                // The following settings are based on https://github.com/grpc/grpc/blob/e35db43c07f27cc13ec061520da1ed185f36abd4/doc/keepalive.md ,
-                // and guidance provided on various github issues for grpc-node. They will enable keepalive pings when a
-                // client connection is idle.
-                'grpc.keepalive_permit_without_calls': 1,
-                'grpc.keepalive_timeout_ms': 1000,
-                'grpc.keepalive_time_ms': 5000,
-              }
+              channelOptions
             ),
           loggerFactory: this.configuration.getLoggerFactory(),
           maxIdleMillis: this.configuration
